requirements.txt
numpy
pandas
scikit-learn
crewai
streamlit
python-dotenv
joblib
